HAB_SW_048: Move the project back to C language. Change compiler to gcc.
After some research the fact that C++ could be linked with C code was discovered.
Because of that, it is wise to move the whole project to C language, as there is
no C++ specific functionality (for instance STL) used.

The python demo code was removed as it does not bring any sense no more.